requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
autopep8==2.0.2
mypy==1.5.1
